amigaDebug: Add 'none' config
Using a config overrides a bunch of important options, esp. CPU type/fpu/mmu, which makes it impossible to configure a tf1260-like config. So, to work around this, we'll just optionally skip quickstart altogether, and rely on the default winuae config to have been saved with the correct settings (or use emuflags, which probably works, and better captures config changes in the source tree, but I haven't tried it).

This is really hacky; there's some type definition for model somewhere I couldn't find, and it wouldn't work with fs-uae on not-Windows, but it works around my specific issues, so there ya go.